Accelerate Functions
This commit speeds up calcualations by quite a bit (Could be 1000x)
. Also TimerOutputs was used to determine performance issues. I
discovered that the interpolate function was called everytime a
value was calcualted from the projectors or pseudopotentials. I've
then created the interpolate function once and put it in the Psp
struct so it's only called once. It's pretty obvious where greater
speedups can be applied, but testing accuracy is now more important
.