Using parameteric types to avoid type instability
This also makes ExternalFromFourier to work with GPU as otherwise
element types of the `map(..) do` cannot be inferred.

The ExternalFromReal and ExternalFromFrouier each stores a function, they
need the type of the function as parameter to avoid type
instability when the function is called when constructing the
TermExternal when the basis is instantiated.