Add refinement test
I tried to compare against forces from an exact linearization which
should be more precise than the forces from the refinement procedure
using Schur's complement. However, they do not seem to be more precise
in the test. The relative error between the linearization forces and the
Schur's complement forces is in fact larger than the relative error
between the Schur's complement forces and the reference forces...

This might be because of the low kgrid and Ecut, but the test should not
take too long to run so increasing these parameters is not an option.